expanded supported functions, made GoIOSensor threadsafe
There now is lock and unlock methods around each GoIOSensor call.  These call the built in GoIO SDK lock and unlock calls.  This wasn't done with java synchronize keyword, because that requires leaving the device unlocked all the time.  And the GoIO SDK asserts sometimes when calls are made and it is unlocked.

git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/sensor/goio-jna@23878 6e01202a-0783-4428-890a-84243c50cc2b